Crate clippy

Modules

approx_const

array_indexing

attrs
checks for attributes

bit_mask

block_in_if_condition

collapsible_if
Checks for if expressions that contain only an if expression.

consts

cyclomatic_complexity
calculate cyclomatic complexity and warn about overly complex functions

derive

entry

eq_op

escape

eta_reduction

identity_op

len_zero

lifetimes

loops

map_clone

matches

methods

minmax

misc

misc_early

mut_mut

mut_reference

mutex_atomic
Checks for uses of Mutex where an atomic value could be used

needless_bool
Checks for needless boolean results of if-else expressions

needless_features
Checks for usage of nightly features that have simple stable equivalents

needless_update

no_effect

open_options

panic

precedence

ptr_arg
Checks for usage of &Vec[_] and &String

ranges

returns

shadow

strings
This LintPass catches both string addition and string addition + assignment

temporary_assignment

transmute

types

unicode

utils

zero_div_zero

Macros

if_let_chain!
Produce a nested chain of if-lets and ifs from the patterns:

Functions

plugin_registrar