Enum clippy::consts::Constant

pub enum Constant {
    Str(String, StrStyle),
    Binary(Rc<Vec<u8>>),
    Char(char),
    Int(ConstInt),
    Float(String, FloatWidth),
    Bool(bool),
    Vec(Vec<Constant>),
    Repeat(Box<Constant>, usize),
    Tuple(Vec<Constant>),
}

A LitKind-like enum to fold constant Exprs into.

Variants

Str

a String "abc"

Binary

a Binary String b"abc"

Char

a single char 'a'

Int

an integer, third argument is whether the value is negated

Float

a float with given type

Bool

true or false

Vec

an array of constants

Repeat

also an array, but with only one constant, repeated N times

Tuple

a tuple of constants

Methods

impl Constant

fn as_float(&self) -> Option<f64>

convert this constant to a f64, if possible