Module clippy::strings

This LintPass catches both string addition and string addition + assignment

Note that since we have two lints where one subsumes the other, we try to disable the subsumed lint unless it has a higher level

Structs

StringAdd

StringLitAsBytes

Statics

STRING_ADD

STRING_ADD_ASSIGN

STRING_LIT_AS_BYTES