Struct clippy::utils::conf::Conf

pub struct Conf {
    pub blacklisted_names: Vec<String>,
    pub cyclomatic_complexity_threshold: u64,
    pub doc_valid_idents: Vec<String>,
    pub too_many_arguments_threshold: u64,
    pub type_complexity_threshold: u64,
    pub max_single_char_names: u64,
}

Type used to store lint configuration.

Fields

blacklisted_names

Lint: BLACKLISTED_NAME. The list of blacklisted names to lint about

cyclomatic_complexity_threshold

Lint: CYCLOMATIC_COMPLEXITY. The maximum cyclomatic complexity a function can have

doc_valid_idents

Lint: DOC_MARKDOWN. The list of words this lint should not consider as identifiers needing ticks

too_many_arguments_threshold

Lint: TOO_MANY_ARGUMENTS. The maximum number of argument a function or method can have

type_complexity_threshold

Lint: TYPE_COMPLEXITY. The maximum complexity a type can have

max_single_char_names

Lint: MANY_SINGLE_CHAR_NAMES. The maximum number of single char bindings a scope may have