Function clippy::utils::method_chain_args

pub fn method_chain_args<'a>(expr: &'a Expr, methods: &[&str]) -> Option<Vec<&'a MethodArgs>>

Match an Expr against a chain of methods, and return the matched Exprs.

For example, if expr represents the .baz() in foo.bar().baz(), matched_method_chain(expr, &["bar", "baz"]) will return a Vec containing the Exprs for .bar() and .baz()