Struct clippy::utils::SpanlessHash

pub struct SpanlessHash<'a, 'tcx: 'a> {
    // some fields omitted
}

Type used to hash an ast element. This is different from the Hash trait on ast types as this trait would consider IDs and spans.

All expressions kind are hashed, but some might have a weaker hash.

Methods

impl<'a, 'tcx: 'a> SpanlessHash<'a, 'tcx>

fn new(cx: &'a LateContext<'a, 'tcx>) -> Self

fn finish(&self) -> u64

fn hash_block(&mut self, b: &Block)

fn hash_expr(&mut self, e: &Expr)

fn hash_exprs(&mut self, e: &[P<Expr>])

fn hash_name(&mut self, n: &Name)

fn hash_path(&mut self, p: &Path)

fn hash_stmt(&mut self, b: &Stmt)